feat(confind): implement proxide-confind — ConFind contact-degree algorithm
Rust/rayon reimplementation of Mosaist mstcondeg.cpp (Grigoryanlab@450816a).
33 tests pass; workspace clippy clean; CI bench-compile check added.

Implements:
- ProximityGrid<T> (f64 spatial index replacing CellList)
- extract_f64_backbone / load_pdb_f64 (single f32→f64 boundary)
- ResidueCache with Phase A rotamer placement, backbone-clash pruning,
  interference and permanent-contact accumulation
- Phase B1/B2/C pipeline: parallel CD enumeration, sequential collProb
  merge, parallel freedom computation
- ConFind struct with cache_all, contacts, interference, bb_interaction,
  crowdedness, freedom public API
- confind CLI binary (--p, --rLib, --o, --sel, --cdcut, --threads)
- Tech debt logged: citations #64, rustdoc #65, workspace clippy #66

Pre-existing workspace clippy errors also fixed (proxide-io, proxide-physics,
proxide-geometry, proxide-wasm, proxide_py, proxide_rs).

+++ crates/proxide-geometry/src/geometry/topology.rs
@@ -46,7 +46,7 @@ pub fn assign_template_hydrogens(
     // Collect matched heavy atoms in this residue: (global_idx, template_name)
     // We need to find which residue these hydrogens belong to.
     // For simplicity, we assume all unmatched_h_indices are in the same residue passed in 'template'.
-    let res_info = processed.residue_info.iter().find(|r| r.res_name == template.name && r.res_id == res_start as i32);
+    let res_info = processed.residue_info.iter().find(|r| r.res_name == template.name && r.res_id == res_start);
     
     if let Some(res_info) = res_info {
         let res_heavy_atoms: Vec<(usize, &str)> = local_to_global
